Add support for showing generic profiles
Previously, this app would only show you your own profile at `/profile`

This changes the route to `/profile/your-username`, and allows showing anyone
else's profile at this route as well

If viewing your own profile, data is loaded immediately from the redux
store, which is cached in localStorage. If viewing someone else's, data
is fetched directly in the component, since no other components use this
data currently and we do not need to pollute the global state with it.