fix: when recovering account, create new key
Problem: after someone recovers their account and Wallet uses that
recovery method's key as the main access key in this new browser, if
they then delete that recovery method, their browser loses the ability
to do anything on their behalf.

Solution: when recovering an account, use the recovery method's key to
generate a new access key, and store that new access key in this
browser's local storage. Only store the recovery method's key in memory
(InMemoryKeyStore) so that if something goes wrong with the operation,
they won't have a transient key stored in their local storage.