fix: Ensure `Account` is instantiated with valid `accountId`
`getValidators` creates an `Account` with the passed `accountId` to call
view methods on the validator contracts. Now that this `Account` is 2FA
aware (#2680), it will also call `account.state()` to check if the
current account has 2FA enabled. This call will fail if the original
`accountId` is `undefined`, which `exAccountId` sometimes is. This PR
updates `getValidators()` call to use `accountId` which is known to
exist.